pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "alembic>=1.18.4",
    "fastapi>=0.135.1",
    "httpx>=0.28.1",
    "psycopg2-binary>=2.9.11",
    "pydantic-settings>=2.13.1",
    "python-multipart>=0.0.22",
    "sqlalchemy>=2.0.48",
    "uvicorn[standard]>=0.42.0",
]
